feat: SplashActivity 적용
