+++ presentation/src/main/AndroidManifest.xml
@@ -13,11 +13,15 @@
         android:theme="@style/Theme.Whereandroid"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".OnBoardingActivity"
-            android:exported="true" >
+            android:name=".view.onboarding.OnBoardingActivity"
+            android:exported="false" />
+        <activity
+            android:name=".view.splash.SplashActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Design.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
