+++ presentation/src/main/kotlin/com/sooum/where_android/view/onboarding/OnBoardingSecondFragment.kt
@@ -1,20 +1,20 @@
-package com.sooum.where_android
+package com.sooum.where_android.view.onboarding
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.sooum.where_android.databinding.FragmentOnBoardingThirdBinding
+import com.sooum.where_android.databinding.FragmentOnBoardingSecondBinding
 
-class OnBoardingThirdFragment: Fragment() {
-    private lateinit var binding : FragmentOnBoardingThirdBinding
+class OnBoardingSecondFragment : Fragment() {
+    private lateinit var binding : FragmentOnBoardingSecondBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentOnBoardingThirdBinding.inflate(inflater, container, false)
+        binding = FragmentOnBoardingSecondBinding.inflate(inflater, container, false)
 
 
         return binding.root
