+++ presentation/src/main/res/layout/fragment_on_boarding_first.xml
@@ -4,8 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".AppInfoFragment1">
+    android:orientation="vertical">
 
     <TextView
         android:layout_width="wrap_content"
