+++ presentation/src/main/res/layout/fragment_on_boarding_second.xml
@@ -3,8 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".AppInfoFragment2">
+    android:orientation="vertical">
 
     <TextView
         android:layout_width="wrap_content"
